support servant 0.11
Since 0.9, servant-server added a built-in way to check query string
parameters. This means that the fairly complex query string handling
code we implemented can be scrapped, and replaced with a simpler
version. `StrictQueryString` will fail with a 400 error if the parameter
is absent, present with no value, or fails to parse. `QueryString`, as
of Servant 0.10 or 0.11, fails with 400 if it fails to parse, but gives
a Maybe that's Nothing if the query string is absent or present with no
value. That seems like it's good enough, so we removed our advanced
query string handling code.

Besides that, calls to `linkURI` were added to change Link values into
URI values.