relativize paths
This fixes a bug in which moving the output files would fail because we
now use a relative path for the temporary directory in which the outputs
are stored.

This patch also obviates the inMusicDir and inDir helper functions,
since we do not need to move into the music dir to perform the move.