fix tar file name problem
Sort of. When performing a youtubeDl, we clamp file names so that they
are at most 99 characters, so that later when we transcode and tar them,
they will have the right size. However, there may be files in the
current library that violate the filename size limit, or we may
introduce new methods for uploading music later that could cause a
violation.

Ultimately, what we need to do is do a bit lower-level in the tar
library and construct Entry values by hand, so that we can control the
filepath inside the archive. Currently the filepath inside the archive
is set to match the filepath on disk because we use the `pack` function.